[#394] Fix import points i18n and Utils, fix paths with alias
