+++ apps/blog/src/app/[locale]/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from 'next/navigation'
-import type { Language } from '@/i18n'
+import type { Language } from '@/i18n/i18n.settings'
 
 export default function BlogPage({ params: { locale } }: { params: { locale: Language } }) {
   return redirect(`/${locale}/posts`)
