+++ apps/blog/src/app/[locale]/posts/_utils/isPostShouldBePickedByLocale.ts
@@ -1,6 +1,6 @@
 import type { BlogPostWithTranslates } from '../_content'
-import type { Language } from '@/i18n'
-import { fallbackLanguage } from '@/i18n'
+import type { Language } from '@/i18n/i18n.settings'
+import { fallbackLanguage } from '@/i18n/i18n.settings'
 
 export const isPostShouldBePickedByLocale = (post: Pick<BlogPostWithTranslates, 'slug' | 'lang' | 'translates'>, locale: Language): boolean => {
   /**
