+++ apps/blog/src/app/[locale]/posts/page.tsx
@@ -1,17 +1,18 @@
 import Link from 'next/link'
 import type { Metadata } from 'next'
+
 import { allBlogPostsWithTranslates, uniqTags } from './_content'
 import { filterBlogPosts } from './_utils/filterBlogPosts'
 import { useTranslation } from '@/i18n'
-import type { Language } from '@/i18n'
 import { SearchInput } from '@/components/SearchInput'
 import { Suspense } from 'react'
-import { formatDate } from '@/utils'
 import { TagsSelector } from '@/components/TagsSelector'
 import { ChipsRow } from '@/components/ChipsRow'
 import { Chip } from '@/components/Chip'
 import { EmptyPlaceholder } from '@/components/PostsList'
 import { PostsListItem } from '@/components/PostsListItem'
+import { formatDate } from '@/utils/formatDate'
+import type { Language } from '@/i18n/i18n.settings'
 
 function SearchLoader() {
   return <div className="rounded-md shadow-sm h-16 border-0 text-gray-900" />
